### 2026-01-04 #9726 @editorialbot
[PRE REVIEW]: MFGLib: A Library for Mean-Field Games
**Submitting author:** @matteosantama (<a href="http://orcid.org/0009-0000-1406-2083">Matteo Santamaria</a>)
**Repository:** https://github.com/radar-research-lab/MFGLib/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.1
