### 2026-01-05 4d80ec9
Add DOIs to citations
