+++ joss/paper.bib
@@ -144,38 +144,51 @@
     url           = { https://arxiv.org/abs/2112.01280 },
 }
 
-@misc{benamou:2019,
-    title           = { An entropy minimization approach to second-order variational mean-field games },
-    author          = { Jean-David Benamou and
-                        Guillaume Carlier and
-                        Simone Di Marino and
-                        Luca Nenna },
-    year            = { 2019 },
-    eprint          = { 1807.09078 },
-    archivePrefix   = { arXiv },
-    primaryClass    = { math.OC },
-    url             = { https://arxiv.org/abs/1807.09078 },
+@article{benamou:2019,
+    author      = { Benamou, Jean-David and Carlier, Guillaume and Di Marino, Simone and Nenna, Luca },
+    title       = { An entropy minimization approach to second-order variational mean-field games },
+    journal     = { Mathematical Models and Methods in Applied Sciences },
+    volume      = { 29 },
+    number      = { 08 },
+    pages       = { 1553-1583 },
+    year        = { 2019 },
+    doi         = { 10.1142/S0218202519500283 },
+    URL         = { https://doi.org/10.1142/S0218202519500283 },
+    eprint      = { https://doi.org/10.1142/S0218202519500283 },
 }
 
-@misc{hu:2024,
-    title           = { {MF-OML}: Online Mean-Field Reinforcement Learning with Occupation Measures for Large Population Games },
-    author          = { Anran Hu and Junzi Zhang },
-    year            = { 2024 },
-    eprint          = { 2405.00282 },
-    archivePrefix   = { arXiv },
-    primaryClass    = { math.OC },
-    url             = { https://arxiv.org/abs/2405.00282 },
+
+@article{hu:2024,
+    author      = { Hu, Anran and Zhang, Junzi },
+    title       = { MF-OML: Online Mean-Field Reinforcement Learning with Occupation Measures for Large Population Games },
+    year        = { 2025 },
+    issue_date  = { Nov 2025 },
+    publisher   = { Springer-Verlag },
+    address     = { Berlin, Heidelberg },
+    volume      = { 92 },
+    number      = { 3 },
+    issn        = { 0095-4616 },
+    url         = { https://doi.org/10.1007/s00245-025-10328-5 },
+    doi         = { 10.1007/s00245-025-10328-5 },
+    journal     = { Appl. Math. Optim. },
+    month       = oct,
+    numpages    = { 54 },
+    keywords    = { Mean-field games, Symmetric N-player games, Nash equilibrium, Occupation measure, Multi-agent reinforcement learning, Operator splitting }
 }
 
-@misc{guo:2023,
-    title           = { MF-OMO: An Optimization Formulation of Mean-Field Games },
-    author          = { Xin Guo and Anran Hu and Junzi Zhang },
-    year            = { 2023 },
-    eprint          = { 2206.09608 },
-    archivePrefix   = { arXiv },
-    primaryClass    = { math.OC },
-    url             = { https://arxiv.org/abs/2206.09608 },
+@article{guo:2023,
+    author      = { Guo, Xin and Hu, Anran and Zhang, Junzi },
+    title       = { MF-OMO: An Optimization Formulation of Mean-Field Games },
+    journal     = { SIAM Journal on Control and Optimization },
+    volume      = { 62 },
+    number      = { 1 },
+    pages       = { 243-270 },
+    year        = { 2024 },
+    doi         = { 10.1137/22M1524084 },
+    URL         = { https://doi.org/10.1137/22M1524084 },
+    eprint      = { https://doi.org/10.1137/22M1524084 },
 }
+
 
 @inproceedings{akiba:2019,
     title       = { Optuna: A next-generation hyperparameter optimization framework },
[…]
