### 2026-01-13 493e9a7
Conform to newly released JOSS submission requirements
