+++ joss/paper.md
@@ -30,60 +30,75 @@ bibliography: paper.bib
 
 # Summary
 
-Mean-field games (MFGs) provide scalable models for large-population strategic interactions. 
-They approximate an $N$-player game by analyzing the limiting regime as $N \to \infty$, 
-replacing explicit multi-agent interactions with the interaction between a representative agent 
-and the population distribution [@lasry:2007]. It has been shown that the Nash equilibrium policy
-of the mean-field game is an $\epsilon$-Nash equilibrium of the $N$-player game with $\epsilon = O(1 / \sqrt{N})$ 
-[@huang:2006] and in practice even games with small $N$ on the order of tens can be well-approximated by MFGs 
-[@guo:2019; @kizilkale:2019; @cabannes:2021]. Due to their tractability, MFG models have become widely used in 
-applications such as digital advertising, high-frequency trading, dynamic pricing, transportation, and behavioral modeling.
+Mean-field games (MFGs) provide scalable models for large-population strategic interactions. They approximate an 
+$N$-player game by analyzing the limiting regime as $N \to \infty$, replacing explicit multi-agent interactions with 
+the interaction between a representative agent and the population distribution [@lasry:2007]. It has been shown that 
+the Nash equilibrium policy of the mean-field game is an $\epsilon$-Nash equilibrium of the $N$-player game with 
+$\epsilon = O(1 / \sqrt{N})$ [@huang:2006] and in practice even games with small $N$ on the order of tens can be 
+well-approximated by MFGs [@guo:2019; @kizilkale:2019; @cabannes:2021]. Due to their tractability, MFG models have 
+become widely used in applications such as digital advertising, high-frequency trading, dynamic pricing, transportation, 
+and behavioral modeling.
 
 Despite the rapid growth of the MFG literature, however, researchers and practitioners lack a unified, 
 open-source software package for defining and solving their own MFG problems. 
 
-**MFGLib** is an open-source Python library that addresses this gap by providing:
+MFGLib is an open-source Python library that addresses this gap by providing:
 
 * A modular and extensible API for defining arbitrary discrete-time finite-state MFGs
 * Implementations of state-of-the-art algorithms for computing (approximate) Nash equilibria
 * A collection of ten customizable benchmark environments drawn from the literature
-* Tight integration with **Optuna** [@akiba:2019] to provide automatic hyperparameter selection
+* Tight integration with Optuna [@akiba:2019] to provide automatic hyperparameter selection
 * Clear documentation and examples, facilitating both research and industry use
 
 The library is implemented in Python, maintained on GitHub, and can be installed via `pip install mfglib`. Full 
 documentation, tutorials, and example notebooks are available at https://mfglib.readthedocs.io/en/latest/.
 
-# Statement of need
+# Statement of Need
 
-Large population games are ubiquitous in real-world problems. As the number of players in the game grows, however, the 
-computational complexity grows exponentially, and it becomes notoriously hard to solve such problems.
+Large-population games, where a massive number of agents interact, are ubiquitous in fields such as economic modeling, 
+crowd dynamics, and smart-grid management. However, as the number of agents $N$ increases, traditional game-theoretic 
+methods face exponential computational growth. MFGs provide a tractable approximation by modeling the collective 
+behavior of agents in the infinite-population limit ($N \to \infty$). Despite the mathematical maturity of MFG theory, 
+the research ecosystem lacks a standardized software framework. This has led to a fragmented landscape where researchers 
+must frequently re-implement population dynamics and equilibrium solvers from scratch for every publication, a process 
+that is time-intensive and prone to implementation errors.
 
-Various libraries have been developed for $N$-player games, such as **QuantEcon** [@batista:2024], **Nashpy**
[…]
