### 2026-01-18 e56e60a
Bump required minimum Python version to 3.11
