+++ mfglib/alg/abc.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import abc
 import json
-import sys
 import time
 import warnings
 from pathlib import Path
@@ -14,15 +13,11 @@ from typing import (
     Iterable,
     Literal,
     Protocol,
+    Self,
     Sequence,
     TypedDict,
     TypeVar,
 )
-
-if sys.version_info >= (3, 11):
-    from typing import Self
-else:
-    from typing_extensions import Self
 
 import optuna
 import torch
