### 2026-01-18 cc15414
Bump required minimum Python version to 3.11 (#61)
