+++ mfglib/alg/occupation_measure_inclusion.py
@@ -1,13 +1,7 @@
 from __future__ import annotations
 
-import sys
 from dataclasses import dataclass
-from typing import Any
-
-if sys.version_info >= (3, 11):
-    from typing import Self
-else:
-    from typing_extensions import Self
+from typing import Any, Self
 
 import numpy as np
 import optuna
