+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.2.0"
 description = "A Library for Mean-Field Games."
 license = { text = "MIT" }
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.11,<3.14"  # osqp doesn't provide wheels for >=3.14 yet
 authors = [
     { name = "Xin Guo", email = "xinguo@berkeley.edu" },
     { name = "Anran Hu", email = "anranhu0107@gmail.com" },
@@ -17,20 +17,17 @@ dependencies = [
     "torch~=2.0",
     "optuna~=3.0",
     "osqp~=0.0",
-    "numpy~=1.0 ; python_version < '3.10'",
-    "numpy~=2.0 ; python_version >= '3.10'",
-    "scipy<1.13 ; python_version < '3.10'",
-    "scipy~=1.13 ; python_version >= '3.10'",
+    "numpy~=2.0",
+    "scipy~=1.17 ",
     "rich~=14.0",
-    "typing-extensions ; python_version < '3.11'"
 ]
 
 [tool.poetry.group.dev.dependencies]
-isort = "~=6.0"
-black = "~=25.0"
+isort = "~=7.0"
+black = "~=26.0"
 mypy = "~=1.0"
 ruff = "~=0.9"
-pytest = "~=8.0"
+pytest = "~=9.0"
 coverage = "~=7.0"
 towncrier = "~=25.0"
 
