+++ .github/workflows/on-pr.yml
@@ -19,16 +19,15 @@ jobs:
           poetry run black . --check
           poetry run isort . --check
           poetry run ruff check mfglib tests
-          poetry run mypy --python-version 3.9
-          poetry run mypy --python-version 3.10
           poetry run mypy --python-version 3.11
           poetry run mypy --python-version 3.12
+          poetry run mypy --python-version 3.13
   test:
     strategy:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+        python-version: ["3.11", "3.12", "3.13"]
     defaults:
       run:
         shell: bash
