### 2026-01-21 #9726 @matteosantama
Hi @danielskatz,

Yes, we are working on it! We hope to have the updated version ready for you sometime next week.
