### 2026-01-26 #9726 @matteosantama
Hi @danielskatz,

Our submission is ready to resume the review process!

I've generated the latest PDF above. I notice the "Download article proof" link results in a 404, though. Is that a problem on our end?
