### 2026-01-30 2fadbf0
Update AI Disclosure to meet requirments
