### 2026-01-30 #9726 @matteosantama
@danielskatz The submission includes all required sections described in the [software paper content](https://joss.readthedocs.io/en/latest/review_criteria.html#software-paper-content) section.

We look forward to getting the review process started!
