### 2026-02-16 #9726 @Nikoleta-v3
@MichalisPanayides @stephenfmann 👋🏻 Would any of you be willing to review this submission for JOSS?

The submission is titled: “MFGLib: A Library for Mean-Field Games.” You can find more details at the top of the GitHub issue for the submission. 🆙

As you already know, we conduct our checklist-driven reviews here in GitHub issues and follow these guidelines:
https://joss.readthedocs.io/en/latest/review_criteria.html 😄
