### 2026-02-16 #9726 @MichalisPanayides
Apologies, but I don't have the time to review this now 👍
