### 2026-03-03 #10137 @Nikoleta-v3
Hey  @stephenfmann, @FranziLesi (@matteosantama) 👋🏻 this is the review thread for the paper. All of our communications will take place here from now on.

As a reviewer, your first step is to create a checklist for your review by entering:

at the top of a new comment in this thread.

These checklists contain the JOSS requirements ✅ As you go through the submission, please check off any items that have been satisfied. The first comment in this thread also includes links to the JOSS reviewer guidelines.

You’re both familiar with the JOSS process, but just as a quick reminder: reviewers are encouraged to submit issues and pull requests directly on the software repository. When doing so, please mention https://github.com/openjournals/joss-reviews/issues/10137 so a link to this thread is created (and I can keep track of everything that’s happening). Alternatively, feel free to leave your reviews as comments on this issue!

We aim to complete reviews within about 2–4 weeks. Please let me know if either of you needs more time. We can also use EditorialBot to set automatic reminders if you’ll be away for a known period.

Feel free to ping me (@Nikoleta-v3) anytime if you have questions or concerns 😄🙋🏻
