### 2026-03-03 #10137 @stephenfmann
Hey, I just have a couple of queries for the author at this stage:

1. The first author on the paper doesn't appear to have any contributions in the GitHub. The "Contribution and authorship" checklist item asks "Does the full list of paper authors seem appropriate and complete?", so it would be good to get confirmation on the role of the paper's first author in development of the software.
2. The latest version in the GitHub and on PyPI is v0.2.0, but the version listed at the top of this thread is v0.1.1. I can't find a version number in the paper itself so perhaps this isn't a problem, but it would be good to ensure consistency wherever possible.
3. In the paper's references, the DOI links work but the arXiv URLs don't work for me. Maybe something about how pandoc handles the url field in the bib file? Could this be fixed?
4. The example in the paper is really good, except that you don't include the plotting function to reproduce the figures. Would you consider including matplotlib as a dependency and extending the example to include a basic plotting function? I can understand if not, but I wanted to suggest it because I think it will help users get up and running.

I'll be testing the functionality in the next week or so and will be back with more then. Thanks!
