### 2026-03-05 d380e80
Fix bib file so arXiv links are properly formatted
