+++ joss/paper.bib
@@ -1,54 +1,54 @@
 @article{lasry:2007,
-    title     = { Mean field games },
-    author    = { Lasry, Jean-Michel and Lions, Pierre-Louis },
-    journal   = { Japanese Journal of Mathematics },
-    volume    = { 2 },
-    number    = { 1 },
-    pages     = { 229--260 },
-    year      = { 2007 },
-    publisher = { Springer }
+    title     = {Mean field games},
+    author    = {Lasry, Jean-Michel and Lions, Pierre-Louis},
+    journal   = {Japanese Journal of Mathematics},
+    volume    = {2},
+    number    = {1},
+    pages     = {229--260},
+    year      = {2007},
+    publisher = {Springer}
 }
 
 @article{perolat:2021,
-    title   = { Scaling up mean field games with online mirror descent },
-    author  = { Perolat, Julien and
+    title   = {Scaling up mean field games with online mirror descent},
+    author  = {Perolat, Julien and
                 Perrin, Sarah and
                 Elie, Romuald and
                 Lauri{\`e}re, Mathieu and
                 Piliouras, Georgios and
                 Geist, Matthieu and
                 Tuyls, Karl and
-                Pietquin, Olivier },
-    journal = { arXiv preprint arXiv:2103.00623 },
-    year    = { 2021 }
+                Pietquin, Olivier},
+    journal = {arXiv preprint arXiv:2103.00623},
+    year    = {2021}
 }
 
 @inproceedings{cui:2021,
-    title        = { Approximately solving mean field games via entropy-regularized deep reinforcement learning },
-    author       = { Cui, Kai and Koeppl, Heinz },
-    booktitle    = { International Conference on Artificial Intelligence and Statistics },
-    pages        = { 1909--1917 },
-    year         = { 2021 },
-    organization = { PMLR }
+    title        = {Approximately solving mean field games via entropy-regularized deep reinforcement learning},
+    author       = {Cui, Kai and Koeppl, Heinz},
+    booktitle    = {International Conference on Artificial Intelligence and Statistics},
+    pages        = {1909--1917},
+    year         = {2021},
+    organization = {PMLR}
 }
 
 @article{perrin:2020,
-    title   = { Fictitious play for mean field games: Continuous time analysis and applications },
-    author  = { Perrin, Sarah and
+    title   = {Fictitious play for mean field games: Continuous time analysis and applications},
+    author  = {Perrin, Sarah and
                 P{\'e}rolat, Julien and
                 Lauri{\`e}re, Mathieu and
                 Geist, Matthieu and
                 Elie, Romuald and
-                Pietquin, Olivier },
-    journal = { Advances in Neural Information Processing Systems },
-    volume  = { 33 },
-    pages   = { 13199--13213 },
-    year    = { 2020 }
+                Pietquin, Olivier},
+    journal = {Advances in Neural Information Processing Systems},
+    volume  = {33},
+    pages   = {13199--13213},
+    year    = {2020}
 }
 
 @article{lanctot:2019,
-    title           = { {OpenSpiel}: A Framework for Reinforcement Learning in Games },
-    author          = { Marc Lanctot and
+    title           = {{OpenSpiel}: A Framework for Reinforcement Learning in Games},
+    author          = {Marc Lanctot and
                         Edward Lockhart and
                         Jean-Baptiste Lespiau and
                         Vinicius Zambaldi and
@@ -74,48 +74,48 @@
                         Thomas Anthony and
                         Edward Hughes and
                         Ivo Danihelka and
[…]
