### 2026-03-05 e965bf8
Add plotting code to example
