### 2026-03-06 #10137 @stephenfmann
I checked basic functionality with a custom game and `OnlineMirrorDescent`. Everything works as I'd expect. I will leave more detailed checking to those who are more familiar with mean field games. Also installation and `pytest` worked fine. So once the submitting author @matteosantama responds to my above queries I'll be happy to tick off the rest of my checklist.
