### 2026-03-09 #10137 @matteosantama
Hi @stephenfmann

Thank you for agreeing to review our submission. Please find our answers to your questions below:

1. The authors are listed in alphabetical order (by last name). The first author actively directed the project’s scope, and made significant contributions in the development of the mathematical framework underpinning the software.
2. Thank you for highlighting this. Indeed, we’d like the most recent version, v2.0.0, to be reviewed. What do you recommend as the best way to update the version of the submission? Does it suffice to simply update the comment at the top of the thread?
3. Again, thank you for pointing this out. I’ve updated the bibtex file so that in the most recently compiled paper, the arXiv links work properly.
4. We agree that including the code used to produce the plot would be beneficial, and we’ve updated the example in the paper to show exactly how we did it.

Please let us know if there are other questions we can answer!
