### 2026-03-10 #10137 @stephenfmann
Thanks!

For updating the version in the submission, I would ask @Nikoleta-v3, as I'm not sure how to change this.

Now that `polars` and `altair` are included in the example, I think they should be somewhere in the requirements in `pyproject.toml`, maybe even just in `dev.dependencies` or  `doc.dependencies`. That's because one of the first things a new user might do is install the package and try to run a version of your example script. Having to manually install polars and altair would be an extra step that would be a bit annoying. (Especially as the import statements come after the relatively long optimisation procedure, so could trigger a runtime error that renders all that processing redundant.)

After these last two bits are done I'm happy to tick off the remaining checklist items.
