### 2026-03-10 #10137 @matteosantama
Hi @stephenfmann

I've updated the plotting example to use `matplotlib`, which we already list as a `doc.dependency`, rather than `polars` and `altair`. In this way, we don't need to introduce additional plotting dependencies.
