### 2026-03-16 #10137 @Nikoleta-v3
Hello both, and thank you for your review and time, @stephenfmann!

Regarding the version mentioned at the top of the issue: I can update it by running the appropriate command, since it has been agreed that version `v2.0.0` being reviewed.

However, we should update the joss branch to reflect software version 2.0.0 (if you haven’t already), so it’s clear to @FranziLesi which version they are reviewing as well.
