### 2026-03-16 #10137 @matteosantama
Great, thanks @FranziLesi!

@Nikoleta-v3, I've confirmed that the `joss` branch points to v2.0.0 (plus commits related to adding the paper to the repository).
