### 2026-03-30 #10137 @Nikoleta-v3
Thank you for your patience @matteosantama! Everything looks good to me. I have opened a pull request to fix some citations, please take a look and, if you agree, feel free to merge it: https://github.com/radar-research-lab/MFGLib/pull/62.

Additionally, could you please:

- Confirm whether we are sticking with v2.0.0 as the version, or if you would prefer to create and use a new release.
- Archive the reviewed software on Zenodo or a similar service (e.g., Figshare or an institutional repository).
  - Please ensure that the archival deposit includes the correct metadata, such as:
  - The correct title (matching the paper title)
  - An accurate author list
  - Optional: ORCIDs for each author

Finally, please share the DOI of the archived version here.
