### 2026-03-30 064b43e
Update metadata now that OSQP provides wheels for Python 3.14
