+++ .github/workflows/release.yml
@@ -15,7 +15,7 @@ jobs:
           ref: ${{ github.event.pull_request.head.sha }}
       - uses: actions/setup-python@v5
         with:
-          python-version: 3.12
+          python-version: "3.13"
       - uses: snok/install-poetry@v1
       - run: |
           poetry --version
