### 2026-03-30 754ff7c
Update metadata now that OSQP provides wheels for Python 3.14 (#63)
