+++ .github/workflows/on-merge-to-main.yml
@@ -11,7 +11,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
+          python-version: "3.13"
       - uses: snok/install-poetry@v1
       - run: |
           poetry --version
