### 2026-03-30 39d28e1
Prepare for v0.3.0 release
