### 2026-03-30 e837e36
fixing citation style and missing doi (#62)
