+++ joss/paper.bib
@@ -6,7 +6,9 @@
     number    = {1},
     pages     = {229--260},
     year      = {2007},
-    publisher = {Springer}
+    publisher = {Springer},
+    doi = {10.1007/s11537-007-0657-8},
+    url = {https://doi.org/10.1007/s11537-007-0657-8},
 }
 
 @article{perolat:2021,
@@ -103,7 +105,7 @@
     number      = {30},
     pages       = {904},
     author      = {Knight, Vincent and Campbell, James},
-    title       = {Nashpy: A Python library for the computation of Nash equilibria},
+    title       = {Nashpy: A {P}ython library for the computation of Nash equilibria},
     journal     = {Journal of Open Source Software},
 }
 
@@ -130,7 +132,7 @@
                     Watkins, Natasha and
                     Yang, Ziyue and
                     Zhang, Hengcheng},
-    title       = {QuantEcon.py: A community based Python library for quantitative economics},
+    title       = {QuantEcon.py: A community based {P}ython library for quantitative economics},
     journal     = {Journal of Open Source Software},
 }
 
