### 2026-03-30 f250999
Prepare for v0.3.0 release (#64)
