+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "mfglib"
-version = "0.2.0"
+version = "0.3.0"
 description = "A Library for Mean-Field Games."
 license = { text = "MIT" }
 readme = "README.md"
