### 2026-03-30 #10137 @matteosantama
Hi @Nikoleta-v3,

Thank you for the PR; it has been merged into the `joss` branch.

Regarding version, I just published `v0.3.0` of the library to PyPI, and we would like to use it for the publication. Functionally, `v0.2.0` and `v0.3.0` are identical. In the latest release, we raised the minimum supported Python version, which motivated a few other minor downstream changes. Python 3.9 reached its EOL several months back, and we wouldn't want to publish software for an unsupported language version.

One can run `git diff joss..v0.3.0` to verify that the only changes are related to CI workflows, type hints, and dependency versioning.

If it is acceptable to you, I will rebase the `joss` branch onto `main` (specifically, the `v0.3.0` tag) and archive that version of the software on Zenodo, etc.
