### 2026-03-31 #10137 @Nikoleta-v3
Thank you for checking, @matteosantama! Looks good to me, so let’s go with version `v0.3.0`.
