### 2026-04-05 #10137 @Nikoleta-v3
@matteosantama did you get a chance to:

- Archive the reviewed software on Zenodo or a similar service (e.g., Figshare or an institutional repository).
  - Please ensure that the archival deposit includes the correct metadata, such as:
  - The correct title (matching the paper title)
  - An accurate author list
  - Optional: ORCIDs for each author

Finally, please share the DOI of the archived version here.
