### 2026-04-13 925d67c
Update paper.bib
