+++ joss/paper.bib
@@ -21,7 +21,8 @@
                 Geist, Matthieu and
                 Tuyls, Karl and
                 Pietquin, Olivier},
-    journal = {arXiv preprint arXiv:2103.00623},
+    journal = {arXiv},
+    doi           = {https://doi.org/10.48550/arXiv.2103.00623},
     year    = {2021}
 }
 
@@ -83,6 +84,7 @@
     primaryClass    = {cs.LG},
     journal         = {CoRR},
     volume          = {abs/1908.09453},
+    doi           = {https://doi.org/10.48550/arXiv.1908.09453},
     url             = {http://arxiv.org/abs/1908.09453},
 }
 
@@ -105,7 +107,7 @@
     number      = {30},
     pages       = {904},
     author      = {Knight, Vincent and Campbell, James},
-    title       = {Nashpy: A {P}ython library for the computation of Nash equilibria},
+    title       = {Nashpy: A {P}ython library for the computation of {N}ash equilibria},
     journal     = {Journal of Open Source Software},
 }
 
@@ -137,12 +139,13 @@
 }
 
 @misc{cui:2022,
-    title         = {Learning Graphon Mean Field Games and Approximate Nash Equilibria},
+    title         = {Learning Graphon Mean Field Games and Approximate {N}ash Equilibria},
     author        = {Kai Cui and Heinz Koeppl},
     year          = {2022},
     eprint        = {2112.01280},
     archivePrefix = {arXiv},
     primaryClass  = {cs.GT},
+    doi           = {https://doi.org/10.48550/arXiv.2112.01280},
     url           = {https://arxiv.org/abs/2112.01280},
 }
 
@@ -246,7 +249,8 @@
                 Bayen, Alexandre M and
                 Goubault, Eric and
                 Elie, Romuald},
-    journal = {arXiv preprint arXiv:2110.11943},
+    journal = {arXiv},
+    doi           = {https://doi.org/10.48550/arXiv.2110.11943},
     year    = {2021}
 }
 
@@ -258,6 +262,7 @@
       archivePrefix = {arXiv},
       primaryClass  = {cs.GT},
       url           = {https://arxiv.org/abs/2307.07911},
+      doi           = {https://doi.org/10.48550/arXiv.2307.07911},
 }
 
 @misc{becherer:2025,
@@ -268,6 +273,7 @@
       archivePrefix = {arXiv},
       primaryClass  = {math.OC},
       url           = {https://arxiv.org/abs/2309.07877},
+      doi           = {https://doi.org/10.48550/arXiv.2309.07877},
 }
 
 @misc{yardim:2024,
@@ -278,6 +284,7 @@
       archivePrefix = {arXiv},
       primaryClass  = {cs.GT},
       url           = {https://arxiv.org/abs/2408.15173},
+      doi           = {https://doi.org/10.48550/arXiv.2408.15173},
 }
 
 @inproceedings{fridovich:2020,
