### 2026-04-13 #10137 @danielskatz
I suggest the changes in https://github.com/radar-research-lab/MFGLib/pull/65 - please merge this, or let me know what you disagree with. Please also then add the country to the third affiliation.
