### 2026-04-13 48432d2
Update paper.md
Add Citadel Securities nationality
