### 2026-04-13 #10137 @danielskatz
👋 @matteosantama - sorry, I created some errors, which are now fixed in https://github.com/radar-research-lab/MFGLib/pull/66
