### 2026-04-13 cabc3ef
fixes in paper.bib (#66)
