+++ joss/paper.bib
@@ -22,7 +22,7 @@
                 Tuyls, Karl and
                 Pietquin, Olivier},
     journal = {arXiv},
-    doi           = {https://doi.org/10.48550/arXiv.2103.00623},
+    doi           = {10.48550/arXiv.2103.00623},
     year    = {2021}
 }
 
@@ -84,7 +84,7 @@
     primaryClass    = {cs.LG},
     journal         = {CoRR},
     volume          = {abs/1908.09453},
-    doi           = {https://doi.org/10.48550/arXiv.1908.09453},
+    doi             = {10.48550/arXiv.1908.09453},
     url             = {http://arxiv.org/abs/1908.09453},
 }
 
@@ -145,7 +145,7 @@
     eprint        = {2112.01280},
     archivePrefix = {arXiv},
     primaryClass  = {cs.GT},
-    doi           = {https://doi.org/10.48550/arXiv.2112.01280},
+    doi           = {10.48550/arXiv.2112.01280},
     url           = {https://arxiv.org/abs/2112.01280},
 }
 
@@ -250,7 +250,7 @@
                 Goubault, Eric and
                 Elie, Romuald},
     journal = {arXiv},
-    doi           = {https://doi.org/10.48550/arXiv.2110.11943},
+    doi     = {10.48550/arXiv.2110.11943},
     year    = {2021}
 }
 
@@ -262,7 +262,7 @@
       archivePrefix = {arXiv},
       primaryClass  = {cs.GT},
       url           = {https://arxiv.org/abs/2307.07911},
-      doi           = {https://doi.org/10.48550/arXiv.2307.07911},
+      doi           = {10.48550/arXiv.2307.07911},
 }
 
 @misc{becherer:2025,
@@ -273,7 +273,7 @@
       archivePrefix = {arXiv},
       primaryClass  = {math.OC},
       url           = {https://arxiv.org/abs/2309.07877},
-      doi           = {https://doi.org/10.48550/arXiv.2309.07877},
+      doi           = {10.48550/arXiv.2309.07877},
 }
 
 @misc{yardim:2024,
@@ -284,7 +284,7 @@
       archivePrefix = {arXiv},
       primaryClass  = {cs.GT},
       url           = {https://arxiv.org/abs/2408.15173},
-      doi           = {https://doi.org/10.48550/arXiv.2408.15173},
+      doi           = {10.48550/arXiv.2408.15173},
 }
 
 @inproceedings{fridovich:2020,
