### 2026-04-13 ba6fbc8
Add JOSS citation information
