### 2026-04-13 d4adccf
Add JOSS citation information (#67)
