### 2026-04-13 #10137 @danielskatz
Congratulations to @matteosantama (Matteo Santamaria) and co-authors on your JOSS publication

JOSS completely depends on volunteers - we couldn't do this without you
